fix: Added separate plugins for each stage of the full test
