+++ test/plugin-tester.test.ts
@@ -4,12 +4,15 @@ import path from 'node:path';
 import { ResourceOperation } from 'codify-schemas/src/types/index.js';
 import deepMatches from 'lodash.matches';
 import differenceWith from 'lodash.differencewith';
+import { PluginProcess } from '../src/plugin-process';
+import * as fs from 'node:fs';
 
+const pluginPath = path.join(__dirname, './test-plugin.ts');
 
 describe('Plugin tester integration tests', () => {
-  let plugin: PluginTester;
+  let plugin: PluginProcess;
   beforeEach(() => {
-    plugin = new PluginTester(path.join(__dirname, './test-plugin.ts'));
+    plugin = new PluginProcess(path.join(__dirname, './test-plugin.ts'));
   })
 
   afterEach(() => {
@@ -120,7 +123,7 @@ describe('Plugin tester integration tests', () => {
 
   it('Has helpers that can test a resource', async () => {
     // No expect needed here. This passes if it doesn't throw.
-    await plugin.fullTest([{
+    await PluginTester.fullTest(pluginPath, [{
       type: 'test',
       propA: 'a',
       propB: 10,
@@ -137,7 +140,7 @@ describe('Plugin tester integration tests', () => {
 
   it('Full test supports plan assertions to ensure the generated plan is correct', async () => {
     // No expect needed here. This passes if it doesn't throw.
-    await plugin.fullTest([{
+    await PluginTester.fullTest(pluginPath, [{
       type: 'test',
       propA: 'a',
       propB: 10,
@@ -167,7 +170,7 @@ describe('Plugin tester integration tests', () => {
 
   it('Full test supports plan assertions to ensure the generated plan is correct (2)', async () => {
     // No expect needed here. This passes if it doesn't throw.
-    await plugin.fullTest([{
+    await PluginTester.fullTest(pluginPath, [{
       type: 'test',
       propA: 'a',
       propB: 10,
@@ -187,7 +190,7 @@ describe('Plugin tester integration tests', () => {
     console.log(differenceWith(['b', 'a'], ['a', 'b', 'c'], (a, b) => deepMatches(a)(b)).length === 0);
 
     // No expect needed here. This passes if it doesn't throw.
-    await plugin.fullTest([{
+    await PluginTester.fullTest(pluginPath, [{
       type: 'test2',
       propB: ['second', 'first'],
       propA: 'a',
@@ -198,7 +201,7 @@ describe('Plugin tester integration tests', () => {
 
   it('Has helpers that can uninstall a resource', async () => {
     // No expect needed here. This passes if it doesn't throw.
-    await plugin.uninstall([{
+    await PluginTester.uninstall(pluginPath, [{
       type: 'test-uninstall',
       propA: 'a',
       propB: 10,
@@ -207,20 +210,18 @@ describe('Plugin tester integration tests', () => {
   })
 
   it('Has helpers that can uninstall a resource (errors out when unsuccessful)', async () => {
-    // const plugin = new PluginTester(path.join(__dirname, './test-plugin.ts'));
-    //
-    // // No expect needed here. This passes if it doesn't throw.
-    // await expect(async () => plugin.uninstall([{
-    //   type: 'test',
-    //   propA: 'a',
-    //   propB: 10,
-    //   propC: 'c',
-    // }])).rejects.toThrowError();
+    // No expect needed here. This passes if it doesn't throw.
+    await expect(async () => PluginTester.uninstall(pluginPath, [{
+      type: 'test',
+      propA: 'a',
+      propB: 10,
+      propC: 'c',
+    }])).rejects.toThrowError();
   })
 
 
   it('Can test modify', { timeout: 50000000 }, async () => {
-    await expect(() => plugin.fullTest([{
+    await expect(() => PluginTester.fullTest(pluginPath, [{
       type: 'test-modify',
       propA: 'a',
       propB: 10,
@@ -236,14 +237,13 @@ describe('Plugin tester integration tests', () => {
     })).rejects.toThrowError();
   })
 
-  it('Will call destory with the correct parameters (modify)', { timeout: 50000000 }, async () => {
-//     const plugin = new PluginTester(path.join(__dirname, './test-plugin.ts'));
-//
-//     await expect(async () => await plugin.fullTest([{
+//   it('Will call destory with the correct parameters (modify)', { timeout: 50000000 }, async () => {
+//     await expect(async () => await PluginTester.fullTest(pluginPath, [{
 //       type: 'test-modify',
 //       propA: 'a',
 //       propB: 10,
 //     }], {
+//       skipUninstall: true,
 //       testModify: {
 //         modifiedConfigs: [{
 //           type: 'test-modify',
@@ -270,54 +270,58 @@ describe('Plugin tester integration tests', () => {
 // }
 // `
 //     )
+//   })
+
+  it('Works when uninstalling two resources', { timeout: 300000 }, async () => {
+    try {
+      await PluginTester.fullTest(pluginPath, [{
+        type: 'test-destroy',
+        propA: 'a',
+        propB: 10,
+      }, {
+        type: 'test-destroy-2',
+        propA: 'a',
+        propB: 20,
+      }], {
+        validateDestroy(plan) {
+          expect(plan.length).to.eq(2);
+          expect(plan[0]).toMatchObject({
+            operation: 'destroy',
+            resourceType: 'test-destroy-2',
+          })
+          expect(plan[1]).toMatchObject({
+            operation: 'destroy',
+            resourceType: 'test-destroy',
+          })
+        }
+      });
+    } finally {
+      try { fs.rmSync('test-destroy'); } catch (e) {}
+      try { fs.rmSync('test-destroy-2') } catch (e) {}
+    }
   })
 
-  it('Works when uninstalling two resources', async () => {
-    await plugin.fullTest([{
-      type: 'test-destroy',
-      propA: 'a',
-      propB: 10,
-    }, {
-      type: 'test-destroy-2',
-      propA: 'a',
-      propB: 20,
-    }], {
-      validateDestroy(plan) {
-        expect(plan.length).to.eq(2);
-        expect(plan[0]).toMatchObject({
-          operation: 'destroy',
-          resourceType: 'test-destroy-2',
-        })
-        expect(plan[1]).toMatchObject({
-          operation: 'destroy',
-          resourceType: 'test-destroy',
-        })
-      }
-    });
-  })
 
-  it('Can uninstall two resources with the same type', async () => {
-    await plugin.fullTest([{
-      type: 'test-destroy',
-      propA: 'a',
-      propB: 10,
-    }, {
-      type: 'test-destroy',
-      propA: 'a',
-      propB: 20,
-    }], {
-      validateDestroy(plan) {
-        expect(plan.length).to.eq(2);
-        expect(plan[0]).toMatchObject({
-          operation: 'destroy',
-          resourceType: 'test-destroy',
-        })
-        expect(plan[1]).toMatchObject({
-          operation: 'destroy',
-          resourceType: 'test-destroy',
-        })
-      }
-    });
-  })
-
+  // it('Can uninstall two resources with the same type', async () => {
+  //   await plugin.fullTest([{
+  //     type: 'test-destroy',
+  //     propA: 'a',
+  //     propB: 10,
+  //   }, {
+  //     type: 'test-destroy',
+  //     propA: 'a',
+  //     propB: 20,
+  //   }], {
+  //     validateDestroy(plan) {
+  //       expect(plan.length).to.eq(2);
+  //       expect(plan[0]).toMatchObject({
+  //         operation: 'destroy',
+  //         resourceType: 'test-destroy',
+  //       })
+  //       expect(plan[1]).toMatchObject({
+  //         operation: 'destroy',
+  //         resourceType: 'test-destroy',
+  //       })
+  //     }
+  //   });
 })
